#116 [Feature-Request] Don't round kg values of pyhsical data - Values edited as float with 2 digit most for each plugin

+++ js/plugins/activities/src/mappers/mapTCXtoFHIR.js
@@ -116,7 +116,7 @@ export default function mapTCXtoFHIR(tcxData) {
                                 text: 'Calories',
                             },
                             valueQuantity: {
-                                value: parseInt(lapChildNode.textContent),
+                                value: parseFloat(lapChildNode.textContent),
                                 unit: 'cal',
                                 system: 'http://unitsofmeasure.org',
                                 code: 'cal',
